refactor slider done
